fix owner populate issue

+++ init/index.js
@@ -1,29 +1,32 @@
 const mongoose = require("mongoose");
-
-const initData = require("./data.js");
 const Listing = require("../models/listing.js");
+const User = require("../models/user.js");
+const initData = require("./data.js");
 
-const MONGO_URL = "mongodb://127.0.0.1:27017/wanderlust"; // ✅ Add this line
-main().then(() => {
-        console.log("connected to DB");
-    })
-    .catch((err) => {
-        console.log(err);
-    });
-
-//databse k liye
 async function main() {
-    await mongoose.connect(MONGO_URL);
+    await mongoose.connect(process.env.DB_URL);
 }
 
-const initDB = async() => {
-    initData.data = initData.data.map((obj) => ({...obj,
-        owner: "688b840fd89eeb50cdf2b944"
+main().then(() => console.log("DB connected"));
+
+const seedDB = async () => {
+
+    const user = await User.findOne(); // 🔥 IMPORTANT
+
+    if (!user) {
+        console.log("No user found, create user first!");
+        return;
+    }
+
+    initData.data = initData.data.map(obj => ({
+        ...obj,
+        owner: user._id
     }));
+
     await Listing.deleteMany({});
     await Listing.insertMany(initData.data);
-    console.log("data was intialized");
-}
-initDB();
 
-app.use(methodOveride("_method"));
+    console.log("Seed done");
+};
+
+seedDB();
